Done with almost every page except for - Categories for each product i.e. Laptops, phones etc - Chats page - Cart -Payment
You will need to follow and build on the structure to complete the website. You should start on it ASAP so that we could discuss in class next wk

P.S. Some of the website styling do not look very nice. If you know how to fix it you can do it